refactor(native): extract sidecar startup
